requirements.dev.txt
-r requirements.txt

watchdog

ipython
ipdb
httpie

black

flake8-builtins
flake8-bugbear
flake8-comprehensions
flake8-pytest-style
pep8-naming
dlint
python-dotenv
